refactor: keep KnownFolder in fsutil_windows.go
